Migrated project dependencies to require.js
Really naive migration for now, but a step toward organizing this thing better.

This migration required switching from a packaged version of ace to the
unpackaged version. There may be a way to use packaged versions of ace with
require.js, but I wasn't able to figure it out. (This way has the definite bonus
of working!)