Added JSON object containing a list of underscore fns
This object should be considered the master metadata document for underscore
functions. We'll include function descriptions, filler text for the editor's
start state, and more in here.

I'll likely want to move this into a database when I have enough data to
justify doing so. (Probably when I implement accounts.)