Made test error display more descriptive
By default, Mocha shows the user the whole stack trace for a failed test.
While that's useful, it's a little too verbose for the limited space we're
dealing with.

Previously I coped with this by removing the error and stack trace entirely. This
commit adds the error message, but excludes the stack trace. A nice middle
ground!