Implemented dynamic underscore function invalidation
On pageload, each player's specrunner will now check against the current function
and its aliases and remap all those functions to undefined. (This prevents users
from simply using aliases to make the tests pass.) It also reassigns those functions
to the contents of the user's editor so we can run it through the test suite.